% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_abs_areas.R
\name{abs_hiearchy}
\alias{abs_hiearchy}
\title{Load ABS geography hierarchy}
\usage{
abs_hiearchy()
}
\value{
Table of region names and IDs
}
\description{
Load ABS geography hierarchy
}
\details{
Useful for checking names of regions
}
\examples{
H <- abs_hierarchy()
H[stringr::str_detect(SA2_NAME_2021, "Franks")]

}
