% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avgLatLon_by.R
\name{avgLatLon_by_POSTCODE}
\alias{avgLatLon_by_POSTCODE}
\title{Summary coordinates by postcode and street}
\usage{
avgLatLon_by_POSTCODE()
}
\value{
A keyed \code{data.table} with columns \code{avgLat} and \code{avgLon}.
}
\description{
Tables with the average latitude and longitude of each postcode,
and each street within each postcode. Weighted by addresses, provided when
exact geocoding cannot identify an address (either because of incomplete or
faulty addresses).
}
