% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data}
\alias{data}
\alias{ADDRESS_DETAIL_ID__by__LATLON}
\alias{STREET_ID_vs_ADDRESS_ID}
\alias{STREET_LOCALITY_ID__STREET_NAME_STREET_TYPE_CODE}
\alias{street_type_decoder}
\title{Datasets used internall for decoding}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 14500797 rows and 3 columns.}
\usage{
ADDRESS_DETAIL_ID__by__LATLON

STREET_ID_vs_ADDRESS_ID

STREET_LOCALITY_ID__STREET_NAME_STREET_TYPE_CODE

street_type_decoder
}
\keyword{data}
\keyword{datasets}
