% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_abs_areas.R
\name{fetch_abs_areas}
\alias{fetch_abs_areas}
\title{Fetch addresses based on abs statistical areas}
\usage{
fetch_abs_areas(areas, area_type = c("MB", "SA1", "SA2", "SA3", "SA4"))
}
\arguments{
\item{areas}{IDs of statistical areas to be returned}

\item{area_type}{The type of statistical area requested. Beware of larger areas}
}
\value{
address table containing all addresses in the specified areas
}
\description{
Fetch addresses based on abs statistical areas
}
\examples{
fetch_abs_areas(109011176, area_type="SA2")
}
