% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchpostcode.R
\name{fetch_postcodes}
\alias{fetch_postcodes}
\title{Fetch entire postcodes}
\usage{
fetch_postcodes(postcodes)
}
\arguments{
\item{postcodes}{one or more postcodes (as numbers, which will be coerced to integers)}
}
\value{
data.table containing all addresses and coordinates
}
\description{
Fetch entire postcodes
}
\details{
Returns an entire postcode of address data. This is useful for generating samples on
a per-postcode basis.
}
\examples{
load(system.file("extdata", "bne_addresses.rda", package = "PSMA"))

bnedata <- fetch_postcodes(unique(bne_addresses$postcode))
# Select 3 addresses at random from each postcode
bnedata[, .SD[sample(.N, 3)], by=.(POSTCODE)]

}
