% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocode.R
\name{geocode}
\alias{geocode}
\title{Geocode address}
\usage{
geocode(flat_number = NULL, number_first = NULL, building_name = NULL,
  street_name = NULL, street_type = NULL, postcode,
  attempt_decode_street_abbrev = FALSE)
}
\arguments{
\item{flat_number, number_first, street_name, street_type, postcode}{The address to be geocoded.}

\item{building_name}{If \code{street_name} is not provided, searches for building names in that
\code{postcode}}

\item{postcode}{Mandatory. Postcode as an integer in which the address is located.}
}
\description{
Geocode address
}
\examples{
geocode(flat_number = NA_character_,
        number_first = 8L,
        street_name = "MALVINA",
        street_type = "PLACE",
        postcode = 3053L)
}
