% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fst.R
\name{get_fst}
\alias{get_fst}
\title{Get data from fst extdata}
\usage{
get_fst(dt = c("ADDRESS_DETAIL_ID__by__LATLON", "STREET_ID_vs_ADDRESS_ID",
  "STREET_LOCALITY_ID__STREET_NAME_STREET_TYPE_CODE"), cache_env = TRUE)
}
\arguments{
\item{dt}{A string, the \code{data.table} to retrieve.}

\item{cache_env}{(logical, default: \code{TRUE}) If \code{FALSE}, the result
is not saved between occurrences. Possibly beneficial if memory is limited.}
}
\value{
The \code{data.table}.
}
\description{
Get data from fst extdata
}
\details{
Purpose is to provide an API to the \code{fst} data
in \code{extdata}. If the \code{\link{PSMA_env}} contains \code{dt}
it is returned; if not, it is retrieved from the \code{fst} file, then
assigned.
}
