% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_geocode.R
\name{mutate_geocode}
\alias{mutate_geocode}
\alias{add_geocode}
\title{Add latitude and longitude columns to a data frame of addresses}
\usage{
mutate_geocode(DT, flat_number = NULL, number_first = NULL,
  building_name = NULL, street_name = NULL, street_type = NULL,
  postcode = NULL, new_names = c("lat", "lon"), overwrite = FALSE)
}
\arguments{
\item{DT}{A \code{data.frame} to which columns will be added.}

\item{flat_number, number_first, building_name, street_name, street_type, postcode}{Columns quoted or unquoted to be passed to \code{\link{geocode}}. If \code{NULL}, \code{DT} must the columns spelled the same as the arguments here.}

\item{new_names}{Character vector of length-2 specifying the new names in the resulting \code{data.frame} for the latitude and longitude respectively.}

\item{overwrite}{If \code{new_names} are present in \code{DT}, should they be overwritten?}
}
\description{
Add latitude and longitude columns to a data frame of addresses
}
