% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMA-options.R
\name{PSMA-options}
\alias{PSMA-options}
\alias{PSMA_env}
\title{PSMA options}
\description{
Options used in the \code{PSMA} package.
}
\details{
\describe{
\item{\code{PSMA_env}}{The environment which \code{fst} files are loaded from and saved into.}
}
}
