% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSMA-package.R
\docType{package}
\name{PSMA-package}
\alias{PSMA}
\alias{PSMA-package}
\title{PSMA: Geocoding and Reverse-Geocoding of Australian Locations}
\description{
Geocoding and reverse-geocoding of Australian addresses using the PSMA data <https://data.gov.au/dataset/geocoded-national-address-file-g-naf>. Incorporates 'G-NAF' by PSMA Australia Limited.
}
\author{
\strong{Maintainer}: Hugh Parsonage \email{hugh.parsonage@gmail.com}

}
\keyword{internal}
