% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revgeocode.R
\name{revgeocode}
\alias{revgeocode}
\title{Match Latitude-Longitude to Australian Address}
\usage{
revgeocode(lat, lon, id = seq_along(lat), close_enough = "10m")
}
\arguments{
\item{lat, lon}{Coordinates. Negative \code{lat} entries correspond to parallels south of the equator.}

\item{id}{An integer vector to identify each \code{lat,lon} coordinate. Defaults to
\code{seq_along(lat)}.}

\item{close_enough}{As in \code{\link[hutilscpp]{match_nrst_haversine}}.}
}
\description{
Match Latitude-Longitude to Australian Address
}
\examples{
revgeocode(-37.8006, 144.9618)




}
