% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-street_type_decoder.R
\docType{data}
\name{street_type_decoder}
\alias{street_type_decoder}
\title{Street type decoder}
\format{
A \code{data.table} of \code{48} rows and \code{2} columns
\describe{
\item{\code{street_abbrev}}{A possible abbreviation}
\item{\code{street_type}}{The standardized/full name}
}
}
\usage{
street_type_decoder
}
\description{
Some street types are commonly abbreviated.
This table decodes them to a standard.
}
\keyword{datasets}
