% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update data}
\usage{
update_data(path = NULL, just_tsv = FALSE)
}
\arguments{
\item{path}{Directory of unzipped folder of the `GNAF` unzipped folder. If
\code{NULL}, the default, an attempt will be made to infer the latest PSMA
url and unzipped in a temporary directory.}

\item{just_tsv}{\code{bool}. If \code{TRUE}, only the .tsv files are produced.
This is intended to be useful only for this package's development.}
}
\value{
Called for its side-effect. When called, data from `path` are incorporated
into the PSMA package folder.
}
\description{
Update data
}
