% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_meshblock_map}
\alias{update_meshblock_map}
\title{Import meshblock hierarchy}
\usage{
update_meshblock_map(mbexcel)
}
\arguments{
\item{mbexcel}{path to excel file from link below}
}
\description{
Import meshblock hierarchy
}
\details{
converts the statistical area hierarchy, not including shapes, to a fst
https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/allocation-files/MB_2021_AUST.xlsx

Should be integrated with the main update function
}
\examples{
\dontrun{
destxl <- file.path(tempdir(), "MB_2021_AUST.xlsx")
download.file("https://www.abs.gov.au/statistics/standards/australian-statistical-geography-standard-asgs-edition-3/jul2021-jun2026/access-and-downloads/allocation-files/MB_2021_AUST.xlsx",
destfile=destxl )
update_meshblock_map(destxl)
}

}
