% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/revgeocode.R
\name{revgeocode}
\alias{revgeocode}
\title{Match Latitude-Longitude to Australian Address}
\usage{
revgeocode(lat, lon, id = seq_along(lat), topn = 1, l_infty_e = 0.25,
  close_enough = "10m")
}
\arguments{
\item{lat, lon}{Coordinates. Negative \code{lat} entries correspond to parallels south of the equator.}

\item{topn}{How many addresses to return for each \code{lat}, \code{lon}.}
}
\description{
Match Latitude-Longitude to Australian Address
}
\examples{
revgeocode(-37.8006, 144.9618)


}
