% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Update data}
\usage{
update_data(path = NULL)
}
\arguments{
\item{path}{Directory of unzipped folder of the `GNAF` unzipped folder.}
}
\description{
Update data
}
